refactor: add todoItem
